requirements.txt
av==10.0.0
brotlipy==0.7.0
ConfigParser==5.3.0
cryptography==3.4.7
cryptography==41.0.1
Cython==0.29.23
dl==0.1.0
docutils==0.17
docutils==0.17
HTMLParser==0.0.2
importlib_metadata==4.11.4
importlib_metadata==3.10.0
importlib_metadata==4.12.0
ipaddr==2.2.0
keyring==22.3.0
lockfile==0.12.2
lxml==4.6.3
mypy_extensions==0.4.3
numpy==1.22.3
numpy==1.20.1
opencv_python==4.6.0.66
opencv_python==4.7.0.72
ordereddict==1.1
Pillow==9.0.1
Pillow==8.2.0
Pillow==9.5.0
protobuf==3.19.4
protobuf==3.20.3
pyOpenSSL==20.0.1
pyOpenSSL==23.2.0
pyOpenSSL==23.2.0
pywin32==302
pywin32==227
PyYAML==5.4.1
PyYAML==6.0
streamlit==1.16.0
streamlit_webrtc==0.45.1
typing_extensions==3.7.4.3
wincertstore==0.2.1
xmlrpclib==1.0.1
zipp==3.8.0
zipp==3.4.1
~pencv_python==4.7.0.68
